# doc/api.yaml
openapi: 3.0.3
servers:
  - url: "http://localhost:3000"
info:
  title: WASAPhoto API
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to 
    WASAPhoto! You can upload your photos directly from your PC, and they will be 
    to everyone following you.
  version: 1.0.0
...
components:
  securitySchemes:
    # Bearer authentication using the user identifier instead of the token
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token containing the userID.
  schemas:
    UserID:
      description: user ID
      type: integer
      example: 1234
    Username:
      description: username
      type: string
      example: Maria
      pattern: '^[a-zA-Z0-9]+$'
      minLength: 3
      maxLength: 16
    PhotoID:
      type: integer
      description: ID of the photo.
      example: 1234
    LikeID:
      description: like ID
      type: integer
      example: 1234
    CommentID:
      description: comment ID
      type: integer
      example: 1234
...
  responses:
    UnauthorizedError:
      description: user ID is missing or invalid
      content:
        application/json:
          schema:
            description: Contains an error message
            type: object
            properties:
              message:
                description: error message
                type: string
                example: Invalid user ID supplied
    InvalidSyntax:
      description: Invalid syntax
      content:
        application/json:
          schema:
            description: Contains an error message
            type: object
            properties:
              message:
                description: error message
                type: string
                example: Invalid username format. Please follow the specified requirements.
    InvalidRequest: 
      description: Invalid request
      content:
        application/json:
          schema:
            description: Contains an error message
            type: object
            properties:
              message:
                description: error message
                type: string
                example: Invalid request
    NotFoundError:
      description: Element not found
      content:
        application/json:
          schema: 
            description: Contains an error message
            type: object
            properties:
              message:
                description: error message
                type: string
                example: Element not found
security:
  - bearerAuth: []
paths:
  /session:
    post:
      tags: ["Login"]
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
         application/json:
          schema:
            description: Contains the username
            type: object
            properties:
              name:
                $ref: '#/components/schemas/Username'
        required: true
        
      responses:
        '201':
          description: User log-in action successful
          content:
           application/json:
             schema:
                $ref: '#/components/schemas/UserID'
                    
        
        '400':
          $ref: '#/components/responses/InvalidSyntax'
...
  /users/{userID}/photos/{photoID}/likes:
    put:
      tags: ["Photos"]
      summary: Adds a like to the specified photo
      description: Users can put likes to all the photos they want
      operationId: likePhoto
      parameters:
        - name: userID
          in: path
          required: true
          description: ID of the user.
          schema:
            $ref: '#/components/schemas/UserID'
        - name: photoID
          in: path
          required: true
          description: ID of the photo you want to like.
          schema:
            $ref: '#/components/schemas/PhotoID'
                    
      responses:
        '201':
          description: Photo liked successfully
          content: 
            application/json:
              schema:
                description: Contains a success message
                type: object
                properties:
                  message:
                    description: OK message
                    type: string
                    example: OK
          
        '400':
          $ref: '#/components/responses/InvalidRequest'
          
        '401': 
          $ref: '#/components/responses/UnauthorizedError'
          
        '404':
          $ref: '#/components/responses/NotFoundError'
  /users/{userID}/photos/{photoID}/likes/{likeID}:
    delete:
      tags: ["Photos"]
      summary: Removes a specific like from the specified photo
      description: Users can remove likes to photos
      operationId: unlikePhoto
      parameters:
        - name: userID
          in: path
          required: true
          description: ID of the user.
          schema:
            $ref: '#/components/schemas/UserID'
        - name: photoID
          in: path
          required: true
          description: ID of the photo you want to unlike.
          schema:
            $ref: '#/components/schemas/PhotoID'
        - name: likeID
          in: path
          required: true
          description: ID of the like you want to remove.
          schema:
            $ref: '#/components/schemas/LikeID'
                    
      responses:
        '201':
          description: Photo unliked successfully
          content: 
            application/json:
              schema:
                description: Contains a success message
                type: object
                properties:
                  message:
                    description: OK message
                    type: string
                    example: OK
          
        '400':
          $ref: '#/components/responses/InvalidRequest'
          
        '401': 
          $ref: '#/components/responses/UnauthorizedError'
          
        '404':
          $ref: '#/components/responses/NotFoundError'
  /users/{userID}/photos/{photoID}/comments:
    post:
      tags: ["Photos"]
      summary: Adds a comment to the specified photo
      description: Users can comment photos published by them and by others
      operationId: commentPhoto
      parameters:
        - name: userID
          in: path
          required: true
          description: ID of the user.
          schema:
            $ref: '#/components/schemas/UserID'
        - name: photoID
          in: path
          required: true
          description: ID of the photo you want to comment.
          schema:
            $ref: '#/components/schemas/PhotoID'
      requestBody:
        description: Your comment.
        required: true
        content:
          application/json:
            schema:
              description: Contains the comment
              type: object
              properties:
                comment:
                  description: comment string
                  type: string
                  pattern: '^.*$'
                  minLength: 3
                  maxLength: 16
                  example: Beautiful Photo!
                    
      responses:
        '201':
          description: Photo commented successfully
          content: 
            application/json:
              schema:
                description: Contains the user and their comment with its ID
                type: object
                properties:
                  authorID: 
                    $ref: '#/components/schemas/UserID'
                  commentID:
                    $ref: '#/components/schemas/CommentID'
                  comment:
                    description: comment string
                    type: string
                    pattern: '^.*$'
                    minLength: 3
                    maxLength: 16
                    example: Beautiful Photo!
          
        '400':
          $ref: '#/components/responses/InvalidRequest'
          
        '401': 
          $ref: '#/components/responses/UnauthorizedError'
          
        '404':
          $ref: '#/components/responses/NotFoundError'
  /users/{userID}/photos/{photoID}/comments/{commentID}:
    delete:
      tags: ["Photos"]
      summary: Removes a specific comment from the specified photo
      description: |-
        Removes the comment with the specified ID under the specified photo. Only authors can remove their comment.
      operationId: uncommentPhoto
      parameters:
        - name: userID
          in: path
          required: true
          description: ID of the user.
          schema:
            $ref: '#/components/schemas/UserID'
        - name: photoID
          in: path
          required: true
          description: ID of the photo where the comment to be removed is located.
          schema:
            $ref: '#/components/schemas/PhotoID'
        - name: commentID
          in: path
          required: true
          description: ID of the comment you want to remove
          schema:
            $ref: '#/components/schemas/CommentID'
            
      responses:
        '200':
          description: Comment deleted successfully
          content: 
            application/json:
              schema:
                description: Contains a success message
                type: object
                properties:
                  message:
                    description: OK message
                    type: string
                    example: OK
          
        '400':
          $ref: '#/components/responses/InvalidRequest'
          
        '401': 
          $ref: '#/components/responses/UnauthorizedError'
          
        '404':
          $ref: '#/components/responses/NotFoundError'
  /users/{userID}/photos/{photoID}:
    delete:
      tags: ["Photos"]
      summary: Deletes a photo
      description: |-
        Removes the specified photo. This operation will also remove all associated 
        likes and comments.
      operationId: deletePhoto
      parameters:
        - name: userID
          in: path
          required: true
          description: ID of the user.
          schema:
            $ref: '#/components/schemas/UserID'
        - name: photoID
          in: path
          required: true
          description: ID of the photo you want to remove.
          schema:
            $ref: '#/components/schemas/PhotoID'
      
      responses:
        '200':
          description: Photo deleted successfully
          content: 
            application/json:
              schema:
                description: Contains a success message
                type: object
                properties:
                  message:
                    description: OK message
                    type: string
                    example: OK
          
        '400':
          $ref: '#/components/responses/InvalidRequest'
          
        '401': 
          $ref: '#/components/responses/UnauthorizedError'
          
        '404':
          $ref: '#/components/responses/NotFoundError'